requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-mptt==0.16.0
djangorestframework==3.15.2
dotenv==0.9.9
h11==0.14.0
httptools==0.6.4
idna==3.10
joblib==1.4.2
kombu==5.5.0
msgpack==1.1.0
narwhals==1.31.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
plotly==6.0.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
